app/src/requirements.txt
click==8.0.3
colorama==0.4.4
Flask==1.1.1
Flask-Cors==3.0.8
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.23
protobuf==3.19.1
python-dotenv==0.19.2
six==1.16.0
Werkzeug==2.0.2
